Fix missing checks in API routes

+++ next/src/app/api/transfer/[transferId]/delete/route.js
@@ -13,7 +13,14 @@ export async function POST(req, { params }) {
 
   const transfer = await Transfer.findOne({ author: auth.user._id, _id: transferId })
 
+  if (!transfer) {
+    return NextResponse.json(resp("transfer not found"), { status: 404 })
+  }
+
   await transfer.deleteOne()
+
+  // Do not await this, it will just lag too much. We assume the deletion succeeds.
+  // We can always delete left over files with a tidy script later.
   workerTransferDelete(transfer.nodeUrl, transfer._id.toString())
 
   return NextResponse.json(resp({}))
